Moved decoupled table component to a new components folder
